updated response type

+++ src/main/java/com/taxicalls/driver/repository/DriverRepository.java
@@ -3,6 +3,6 @@ package com.taxicalls.driver.repository;
 import com.taxicalls.driver.model.Driver;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DriverRepository extends CrudRepository<Driver, Integer> {
+public interface DriverRepository extends CrudRepository<Driver, Long> {
 
 }
